Fixed various bugs + better user account type integration
Update company.html, search.js, and general.html templates, and CompanyController.java and SearchAPI.java

Update company.html template to fix a typo in the th:replace attribute.

Update search.js to use the username property of the user object and add the account_type property to the new_user object.

Update general.html template to display icons based on the account_type of the user.

Update CompanyController.java to remove unnecessary logging statement and fix a variable name.

Update SearchAPI.java to modify the SearchResponse class to include the account_type property in the UserSearchResponse class. Modify the search method to filter users and games based on the query.